complite the timer

+++ src/driver/vga.c
@@ -35,7 +35,7 @@ void vga_print(const char *str) {
         return;
     }
 
-    for (uint64_t i = 0; str[i] != '\0'; i++) {
+    for (uint32_t i = 0; str[i] != '\0'; i++) {
         if (str[i] == '\n') {
             uint8_t y = vga_get_cursor_y() + 1;
             vga_set_cursor(0, y);
